refactor(tests): Clearly divide streaming from complete
