+++ src/bytes/mod.rs
@@ -2,5 +2,3 @@
 
 pub mod complete;
 pub mod streaming;
-#[cfg(test)]
-mod tests;
