+++ src/character/mod.rs
@@ -2,8 +2,5 @@
 //!
 //! Functions recognizing specific characters
 
-#[cfg(test)]
-mod tests;
-
 pub mod complete;
 pub mod streaming;
